+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added 
+### Added
 - `MLBMatchupComparison` for fox sports json response for probable starting pitchers
 - `MLBProbableStartingPitcher` fox sports data model for probable starting pitcher
 - Scraping fox sports MLB probable starting pitcher using `MLBProbableStartingPitcherScraper`