+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MLBMatchupComparison` for fox sports json response for probable starting pitchers (#73)
 - `MLBProbableStartingPitcher` fox sports data model for probable starting pitcher (#73)
 - Scraping fox sports MLB probable starting pitcher using `MLBProbableStartingPitcherScraper` (#73)
+### Fixed
+- Ignore baseball-reference MLB all-star matchups (#74)
 
 ## [0.8.2] - 2025-05-17
 ### Fixed